use const originKey
