+++ openapi3/info.go
@@ -63,7 +63,7 @@ func (info *Info) UnmarshalJSON(data []byte) error {
 		return unmarshalError(err)
 	}
 	_ = json.Unmarshal(data, &x.Extensions)
-	delete(x.Extensions, "origin")
+	delete(x.Extensions, originKey)
 	delete(x.Extensions, "title")
 	delete(x.Extensions, "description")
 	delete(x.Extensions, "termsOfService")
